Add support for napari 0.4.19
I previously relied on napari bugs. For example
```self.pointer.data = []```  previously did not cause the data
event handler to be called. Now it does, and this causes recursion.

The update also caused the edge_color event to be called when new points
are added. This caused a problem where occasionally the counter GUI would
be updated when there were extra points added (due to the hack were a
point is added and immediately removed to ensure that no points be selected)

There were other rare bugs related to extra entries being added to the
undo stack, which was solved by ensuring only `added` or `adding` events
cause points to be added, and not both.

More tests are necessary before a napari-3d-counter 0.2.0 release.