loading points populates existing layers
This change makes it so that if you load points from a csv file that has
a cell type with same name as an existing celltype, and the existing
cell type has no points. the points from the csv file applied to the
existing celltype.

The previous behavior was to create a new channel with a [1] suffix that
does not keep any of the style information from the existing layer.

This change is necessary since it allows the 3d counter state to be
fully recovered: style information is stored in the python config and
the points are saved in the csv file.